Moving rework into main file. Ready for pull request.
